Update link to video
The video of this talk is no longer at the original location. However, I
found it at rubyevents.org today. So, I thought I'd fix the link on the
off chance that someone may be interested in watching and listening to
the talk, which is now ten years old. I think it still holds many true
lessons.